feat : 구현을 테스트하는것이 아닌 역할과 인터페이스에 의존하도록 수정
